Use publication date instead of creation date for songs
The grilo tracker2 plugin maps the nie:contentCreated property (the
date of a song) to the CREATION_DATE key.
However, the CREATION_DATE key is supposed to be used to describe the
creation of a media such as an image.

So, in the tracker3 plugin, the nie:contentCreated property now also
maps the PUBLICATION_DATE key (CREATION_DATE is kept for backward
compatibility). So Music, can now use the PUBLICATION_DATE key instead
of the CREATION_DATE one.